% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_analysis_n.R
\name{design_analysis_n}
\alias{design_analysis_n}
\title{Design Analysis: required sample size}
\usage{
design_analysis_n(
  d,
  power,
  sig.level = 0.05,
  rangen = c(2, 1000),
  B = 10000,
  tol = 0.005
)
}
\arguments{
\item{d}{= plausible effect size (Cohen's \eqn{d})}

\item{power}{= required power of the test}

\item{sig.level}{= significance level}

\item{rangen}{= range of \eqn{n} (per group)}

\item{B}{= number of replicates}

\item{tol}{= tolerance}
}
\description{
Internal function. In a comparison between two independent groups,
given a plausible value of effect size (Cohen's \eqn{d}),
the function estimates the sample size (per group) needed to reach a required \code{power} level.
}
\note{
Internal Function, called from function \code{\link{design_analysis}}.
}
\author{
Massimiliano Pastore, Gianmarco Altoè
}
