% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_est.R
\name{design_est}
\alias{design_est}
\title{Retrospective Design Analysis}
\usage{
design_est(
  n1 = 5,
  n2 = n1,
  target_d = NULL,
  target_d_limits = NULL,
  distribution = c("uniform", "normal"),
  k = 1/6,
  sig.level = 0.05,
  B = 500,
  B0 = 500,
  return_data = FALSE
)
}
\arguments{
\item{n1}{= group 1 sample size}

\item{n2}{= group 2 sample size (default = n1)}

\item{target_d}{= plausible effect size (Cohen's \eqn{d})}

\item{target_d_limits}{= vector with lower and upper bounds of plausible effect size (Cohen's \eqn{d})}

\item{distribution}{= probability distribution function \code{uniform} or (truncated) \code{normal}}

\item{k}{= constant for defining the standard deviation of truncated normal}

\item{sig.level}{= significance level}

\item{B}{= number of replicates}

\item{B0}{= number of replicates for simulating \eqn{d} calling \code{\link{sim_estimate}}}

\item{return_data}{= logical, by default is FALSE}
}
\value{
It returns two lists:

\item{call }{reporting input values: \code{n1}, \code{n2}, \code{target_d}, \code{target_d_limits} e \code{distribution}.}

\item{results }{reporting output values: \code{sim_d} (mean of simulated \eqn{d}, see details), \code{est_d} (estimated \eqn{d}), \code{power}, \code{typeS} e \code{typeM}.}

If \code{return_data = TRUE}, a \code{data.frame} (\code{B0}  rows and 3 columns)
is returned with complete information for each sampled value of \eqn{d}:
 \code{power}, \code{typeS} and \code{typeM}.
}
\description{
Performs retrospective design analysis for the comparison between two independent groups.
It estimates power, type S and type M errors given the sample size of each group (\code{n1} and
\code{n2}) and a plausible effect size (Cohen's \eqn{d}). The plausible effect size can be defined as a single value
(\code{target_d}) or as a range of possible values. In the second case, the interval of possible values
(\code{target_d_limits}) as well as the type of distribution (\code{distribution}) can be specified.
Distributions available are uniform (default) or doubly truncated normal.
}
\details{
Default settings consider equal size for the two groups (\code{n1} = \code{n2}).

When a plausible effect size is defined as a range of possible values, the  function \code{\link{sampling_d}}
 is used to sample \code{B0} values of \eqn{d} from the specified distribution, “\code{uniform}”, by default,
 or (doubly truncated) “\code{normal}”. Power, type S and type M are computed for each sampled value of \eqn{d} but only
 mean values are reported in the output. To obtain results for each sampled value of \eqn{d}, set option
 \code{return_data = TRUE}.
}
\examples{
design_est( target_d = .25 ) # default sample size is 5 for both groups
design_est( n1 = 100, target_d = .25 )
design_est( n1 = 100, n2 = 50, target_d = .25 ) # different sample sizes

### Dealing with uncertainty on d ###
design_est( n1 = 100, target_d_limits = c( .2, .3 ) ) # uniform distribution
design_est( n1 = 100, target_d_limits = c( .2, .3 ), distribution = "normal" ) # normal distribution

### Get complete information for each sampled value of d  ###
out <- design_est(n1=50, n2=48, distribution="normal",
                  target_d_limits = c(.20,.30), return_data=TRUE)
out

#### To draw the distribution of simulated power, Type S and Type M errors:  ####
hist(out$results$data$power,main="",xlab="power") # power
hist(out$results$data$typeS,main="",xlab="typeS") # Type S error
hist(out$results$data$typeM,main="",xlab="typeM") # Type M error
}
\author{
Massimiliano Pastore, Gianmarco Altoè
}
