% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_bayes.R
\name{effect_bayes}
\alias{effect_bayes}
\title{Compute the Bayes Factor for Mean Differences}
\usage{
effect_bayes(
  x1,
  x2,
  prior_limits = c(0, 1),
  prior_density = function(y) dcauchy(y, location = 0, scale = 1/sqrt(2))
)
}
\arguments{
\item{x1}{numerical vector with the observed values of the first group.}

\item{x2}{numerical vector with the observed values of the second group.}

\item{prior_limits}{numerical vector with lower and upper bounds of the prior distribution of the mean.}

\item{prior_density}{function, density of the prior distribution of the effect.}
}
\value{
The function returns a list with the following objects: \code{BF} (Bayes factor) and
\code{difference} (difference between group means).
}
\description{
The function computes the Bayes factor to compare two models for the difference between means:
the null model where the difference is zero, and the alternative model where the difference
follows a given prior distribution.
}
\details{
The Bayes factor is computed as the ratio of the posterior probabilities of the alternative
model and the null model.

The domain of the prior distribution should contain zero.
}
\examples{
x1 <- c(0.43, -0.56, -0.39, -0.21, -0.58,  0.65)
x2 <- c(2.38, -0.96, -0.82,  0.28, -0.40,  0.34,  1.82)
effect_bayes(x1, x2, prior_limits=c(0,1), prior_density=function(y) dcauchy(y, location=0, scale=1/sqrt(2)))
}
\author{
Massimiliano Pastore
}
