% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_sd.R
\name{pool_sd}
\alias{pool_sd}
\title{Pooled Standard Deviation}
\usage{
pool_sd(x1, x2, m1, m2, df)
}
\arguments{
\item{x1}{numerical vector with the observed values of the first group.}

\item{x2}{numerical vector with the observed values of the second group.}

\item{m1}{mean of the first group.}

\item{m2}{mean of the second group.}

\item{df}{degrees of freedom.}
}
\value{
The function returns the pooled standard deviation.
}
\description{
Internal function. It computes the pooled standard deviation of two groups.
}
\details{
The degrees of freedom \code{df} are given by the sum of the two group
sizes minus 2.
}
\examples{
x1 <- c(0.43, -0.56, -0.39, -0.21, -0.58,  0.65)
x2 <- c(2.38, -0.96, -0.82,  0.28, -0.40,  0.34,  1.82)
m1 <- mean(x1, na.rm=TRUE)
m2 <- mean(x2, na.rm=TRUE)
df <- length(na.omit(x1)) + length(na.omit(x2)) - 2
pool_sd(x1, x2, m1, m2, df)
}
\author{
Massimiliano Pastore
}
