% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_e.R
\name{sampling_e}
\alias{sampling_e}
\title{Sampling Effect}
\usage{
sampling_e(
  limits = c(0, 1),
  distribution = c("uniform", "truncnorm"),
  h = 1/6,
  B0 = 10000
)
}
\arguments{
\item{limits}{vector with lower and upper bounds of the distribution.}

\item{distribution}{a character string specifying the sampling distribution, must be one of \code{"uniform"} or \code{"truncnorm"}.}

\item{h}{constant defining the standard deviation of the truncated normal distribution.}

\item{B0}{number of replicates.}
}
\value{
The function returns a list with the following objrcts:
\code{mean} (mean of the distribution), \code{sd} (standard deviation of the distribution)
and \code{effect} (vector containing the sampled effect values).
}
\description{
Internal function. It samples \code{B0} values of effect  size from a given bounded
distribution (uniform or doubly truncated normal).
}
\details{
The uniform distribution is defined considering minimum and maximum according to
\code{limits} values. The doubly truncated normal distribution is defined with mean
\eqn{(a+b)/2} and standard deviation \eqn{(b-a)h},
where \eqn{a} and \eqn{b} are respectively the lower and upper values
defined in \code{limits}, and \eqn{h} is a chosen constant.
}
\note{
The \pkg{truncnorm} package is required.
}
\examples{
e <- sampling_e(limits=c(0.2,0.3), distribution="truncnorm", h=1/10, B0=100)
hist(e$effect)
}
\author{
Massimiliano Pastore, Gianmarco Altoè
}
