% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimate.R
\name{sim_estimate}
\alias{sim_estimate}
\title{Simulation of Power and Errors}
\usage{
sim_estimate(
  sim_eff,
  target_eff = sim_eff,
  type = c("cohen", "hedges", "glass", "bayes"),
  n1,
  n2 = n1,
  sig.level = 0.05,
  prior_limits = c(0, 1),
  prior_density = function(y) dcauchy(y, location = 0, scale = 1/sqrt(2)),
  B = 10000,
  return_data = FALSE
)
}
\arguments{
\item{sim_eff}{simulated effect size}

\item{target_eff}{target effect size}

\item{type}{a character string specifying the effect type, must be one of \code{"cohen"},
\code{"hedges"}, \code{"glass"} or \code{"bayes"}.}

\item{n1}{sample size of the first group.}

\item{n2}{sample size of the second group.}

\item{sig.level}{significance level.}

\item{prior_limits}{vector with lower and upper bounds of the prior distribution of the mean.}

\item{prior_density}{function, density of the prior distribution of the mean.}

\item{B}{number of replicates.}

\item{return_data}{logical, if TRUE the function returns also the simulated data.}
}
\value{
Returns a list with the computed values of
\code{power}, \code{typeS}, \code{typeM}, \code{typeN} and \code{typeI}  errors and, if \code{return_data}
is TRUE, a data frame \code{sim_data} with the simulated data.
}
\description{
Internal function. In a comparison between two independent groups,
the function computes through \code{B} simulations the expected power,
Type-S, Type-M, Type-N and Type-I errors, given a plausible effect size (\code{target_eff}) and the sample size of each group.
}
\details{
Type-S error is the probabiliyy that a significant effect is estimated in the wrong direction.

Type-M error is the average overestimation of an effect that emerges as significant.

Type-N error is the probability of false negatives (1-power in a non-Bayesian framework).

Type-I error is the probability of an indecisive outcome when the effect is significative (zero
in a non-Bayesian framework).
}
\note{
Called from \code{\link{design_est}}. For \code{design_analysis}: \code{target_eff} is equal to \code{sim_eff}.
}
\examples{
sim_estimate(sim_eff=0.9, type="cohen", n1=10, n2=15, sig.level=0.05, B=10, return.data=FALSE)
}
\author{
Massimiliano Pastore, Gianmarco Altoè
}
